requirements.txt
pydantic>=2.0,<3.0
pytest>=8.0,<10.0
python-dotenv>=1.0
google-generativeai>=0.5.0